refactor(app): extract locale_settings + currency_formatter to core_app_shell
First slice of the legacy-coins extraction plan (#1111): LocaleSettings
and the currency/date formatters are cross-app user-preference
infrastructure (consumed by coins, money, bill_split, agent chat, and
the auth profile), not coins-local code — and their 12 imports were the
single biggest coupling holding app/lib/features/coins to app
internals.

Same shim pattern as #1101/#1104: files git-mv'd into
packages/core_app_shell/lib/src/, one-line re-export shims left at the
old app/lib/core/utils paths so relative-import consumers are
untouched; test/integration files using package:airo_app absolute
imports switched to package:core_app_shell.

core_app_shell gains core_data + intl dependencies (module.yaml
updated accordingly).
