+++ app/integration_test/coins_user_journeys_test.dart
@@ -1,3 +1,2 @@
 // Deprecated integration test wrapper – replaced with empty stub to avoid duplicate binding.
 void main() {}
-
