+++ app/lib/core/utils/locale_settings.dart
@@ -1,213 +1,1 @@
-import 'dart:convert';
-import 'package:core_data/core_data.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_riverpod/legacy.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-import 'currency_formatter.dart';
-
-/// User locale settings for the app
-class LocaleSettings {
-  final String locale;
-  final String currency;
-  final String dateFormat;
-  final String timeFormat;
-  final String numberFormat;
-
-  const LocaleSettings({
-    this.locale = 'en_IN',
-    this.currency = 'INR',
-    this.dateFormat = 'dd/MM/yyyy',
-    this.timeFormat = 'HH:mm',
-    this.numberFormat = '#,##,##0.00', // Indian numbering system
-  });
-
-  /// Default settings for India
-  static const LocaleSettings india = LocaleSettings();
-
-  /// Default settings for US
-  static const LocaleSettings us = LocaleSettings(
-    locale: 'en_US',
-    currency: 'USD',
-    dateFormat: 'MM/dd/yyyy',
-    numberFormat: '#,##0.00',
-  );
-
-  /// Get currency formatter for current settings
-  CurrencyFormatter get currencyFormatter =>
-      CurrencyFormatter.fromCode(currency);
-
-  /// Get supported currency enum
-  SupportedCurrency get supportedCurrency =>
-      SupportedCurrency.fromCode(currency);
-
-  LocaleSettings copyWith({
-    String? locale,
-    String? currency,
-    String? dateFormat,
-    String? timeFormat,
-    String? numberFormat,
-  }) {
-    return LocaleSettings(
-      locale: locale ?? this.locale,
-      currency: currency ?? this.currency,
-      dateFormat: dateFormat ?? this.dateFormat,
-      timeFormat: timeFormat ?? this.timeFormat,
-      numberFormat: numberFormat ?? this.numberFormat,
-    );
-  }
-
-  Map<String, dynamic> toJson() {
-    return {
-      'locale': locale,
-      'currency': currency,
-      'dateFormat': dateFormat,
-      'timeFormat': timeFormat,
-      'numberFormat': numberFormat,
-    };
-  }
-
-  factory LocaleSettings.fromJson(Map<String, dynamic> json) {
-    return LocaleSettings(
-      locale: json['locale'] as String? ?? 'en_IN',
-      currency: json['currency'] as String? ?? 'INR',
-      dateFormat: json['dateFormat'] as String? ?? 'dd/MM/yyyy',
-      timeFormat: json['timeFormat'] as String? ?? 'HH:mm',
-      numberFormat: json['numberFormat'] as String? ?? '#,##,##0.00',
-    );
-  }
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) return true;
-    return other is LocaleSettings &&
-        other.locale == locale &&
-        other.currency == currency &&
-        other.dateFormat == dateFormat &&
-        other.timeFormat == timeFormat &&
-        other.numberFormat == numberFormat;
-  }
-
-  @override
-  int get hashCode =>
-      Object.hash(locale, currency, dateFormat, timeFormat, numberFormat);
-}
-
-/// Provider for locale settings
-class LocaleSettingsNotifier extends StateNotifier<LocaleSettings> {
-  static const String _storageKey = 'airo_locale_settings';
-  static const String _profileKey = 'airo_user_profile';
-  static const String _currentUserIdKey = 'airo_current_user_id';
-  final int maxPreferenceValueBytes;
-  SharedPreferences? _prefs;
-  KeyValueStore? _store;
-
-  LocaleSettingsNotifier({
-    this.maxPreferenceValueBytes = kKeyValueStorePreferenceMaxValueBytes,
-  }) : super(LocaleSettings.india) {
-    _initialize();
-  }
-
-  Future<KeyValueStore> get _keyValueStore async {
-    final prefs = _prefs ?? await SharedPreferences.getInstance();
-    _prefs = prefs;
-    _store ??= PreferencesStore(prefs, maxValueBytes: maxPreferenceValueBytes);
-    return _store!;
-  }
-
-  Future<void> _initialize() async {
-    try {
-      _prefs = await SharedPreferences.getInstance();
-      final profileSettings = _loadCurrentProfileLocaleSettings(_prefs!);
-      if (profileSettings != null) {
-        state = profileSettings;
-        await _persistString(_storageKey, jsonEncode(profileSettings.toJson()));
-        return;
-      }
-
-      final jsonStr = _prefs?.getString(_storageKey);
-      if (jsonStr != null) {
-        state = LocaleSettings.fromJson(jsonDecode(jsonStr));
-      }
-    } catch (e) {
-      debugPrint('Error loading locale settings: $e');
-      // Keep default India settings on error
-    }
-  }
-
-  Future<void> updateSettings(LocaleSettings settings) async {
-    state = settings;
-    await _persistString(_storageKey, jsonEncode(settings.toJson()));
-    await _persistCurrentProfileSettings(settings);
-  }
-
-  Future<void> setCurrency(String currency) async {
-    await updateSettings(state.copyWith(currency: currency));
-  }
-
-  Future<void> setLocale(String locale) async {
-    await updateSettings(state.copyWith(locale: locale));
-  }
-
-  Future<void> resetToDefaults() async {
-    await updateSettings(LocaleSettings.india);
-  }
-
-  LocaleSettings? _loadCurrentProfileLocaleSettings(SharedPreferences prefs) {
-    final userId = prefs.getString(_currentUserIdKey) ?? 'default';
-    final profileJson = prefs.getString('${_profileKey}_$userId');
-    if (profileJson == null) return null;
-
-    final decoded = jsonDecode(profileJson);
-    if (decoded is! Map<String, dynamic>) return null;
-
-    final localeJson = decoded['localeSettings'];
-    if (localeJson is! Map<String, dynamic>) return null;
-
-    return LocaleSettings.fromJson(localeJson);
-  }
-
-  Future<void> _persistCurrentProfileSettings(LocaleSettings settings) async {
-    final prefs = _prefs;
-    if (prefs == null) return;
-
-    final userId = prefs.getString(_currentUserIdKey) ?? 'default';
-    final key = '${_profileKey}_$userId';
-    final profileJson = prefs.getString(key);
-    if (profileJson == null) return;
-
-    try {
-      final decoded = jsonDecode(profileJson);
-      if (decoded is! Map<String, dynamic>) return;
-
-      final updatedProfile = Map<String, dynamic>.from(decoded)
-        ..['localeSettings'] = settings.toJson()
-        ..['updatedAt'] = DateTime.now().toIso8601String();
-
-      await _persistString(key, jsonEncode(updatedProfile));
-    } catch (e) {
-      debugPrint('Error syncing locale settings to profile: $e');
-    }
-  }
-
-  Future<void> _persistString(String key, String value) async {
-    try {
-      final store = await _keyValueStore;
-      await store.setString(key, value);
-    } on KeyValueStoreValueTooLargeException catch (e) {
-      debugPrint('Locale settings exceeded preference tier: $e');
-    }
-  }
-}
-
-/// Global locale settings provider
-final localeSettingsProvider =
-    StateNotifierProvider<LocaleSettingsNotifier, LocaleSettings>((ref) {
-      return LocaleSettingsNotifier();
-    });
-
-/// Currency formatter provider (derived from locale settings)
-final currencyFormatterProvider = Provider<CurrencyFormatter>((ref) {
-  final settings = ref.watch(localeSettingsProvider);
-  return settings.currencyFormatter;
-});
+export 'package:core_app_shell/src/locale_settings.dart';
