+++ app/test/core/utils/locale_settings_test.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:airo_app/core/utils/locale_settings.dart';
+import 'package:core_app_shell/core_app_shell.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
