+++ packages/core_app_shell/pubspec.yaml
@@ -12,11 +12,14 @@ environment:
   flutter: ">=1.17.0"
 
 dependencies:
+  core_data:
+    path: ../core_data
   core_ui:
     path: ../core_ui
   flutter:
     sdk: flutter
   flutter_riverpod: ^3.3.2
+  intl: ^0.20.2
   shared_preferences: ^2.5.5
 
 dev_dependencies:
